+++ source/blender/draw/engines/eevee/shaders/lightprobe_filter_glossy_frag.glsl
@@ -31,8 +31,8 @@ void main()
   /* Integrating Envmap */
   float weight = 0.0;
   vec3 out_radiance = vec3(0.0);
-  for (float i = 0; i < sampleCount; i++) {
-    vec3 Xi = rand2d_to_cylinder(hammersley_2d(i, sampleCount));
+  for (int i = 0; i < int(sampleCount); i++) {
+    vec3 Xi = rand2d_to_cylinder(hammersley_2d(float(i), sampleCount));
 
     float pdf;
     /* Microfacet normal */
